Done some changes

+++ Aerothon/Models/Entities/Waypoint.cs
@@ -2,8 +2,8 @@
 {
     public class Waypoint
     {
-        public double Lattitude { get; set; }
-        public double Longitude { get; set; }
+        public float Lattitude { get; set; }
+        public float Longitude { get; set; }
         public bool Weather { get; set; }
     }
 }
